pyproject.toml
[project]
name = "subtrans"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.61.0",
    "groq>=1.0.0",
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
